Reorganized files and added randomized inserts (kinda)
